+++ src/components/atoms/FacebookButtons/types.ts
@@ -1,0 +1,10 @@
+import { CommonProps, MoBrixUiComponent } from "../../../utils/global";
+
+export type FacebookButtonsProps = CommonProps & {
+  pageId: string;
+  share?: boolean;
+  small?: boolean;
+  width?: number;
+};
+
+export type FacebookButtonsComponent = MoBrixUiComponent<FacebookButtonsProps>;