+++ playground/public/docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## Versions
 
+- [1.2.0](#120)
 - [1.1.0](#110)
 - [1.0.0](#100)
 
@@ -13,10 +14,18 @@
 
 <br>
 
+### 1.2.0
+
+- `Carousel` arrow fix
+- `ExpandableContainer` arrow fix
+- `FacebookButtons` component
+
+<br>
+
 ### 1.1.0
 
 - `readonly` support for `Rater` component
-- `Review` components
+- `Review` component
 - Clean-up
 
 <br>